pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-toolbox"
version = "0.1.0"
description = "Standardized Agent Runtime — preset-based AI agent gateway with stable HTTP API"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Agent Toolbox" }]
keywords = ["agent", "llm", "gateway", "fastapi", "preset"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn[standard]>=0.29.0,<1.0.0",
    "pydantic>=2.6.0,<3.0.0",
    "jsonschema>=4.21.0,<5.0.0",
    "PyYAML>=6.0.0,<7.0.0",
    "httpx>=0.27.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<10.0.0",
]

[project.scripts]
agent-toolbox = "app.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.setuptools.package-data]
app = ["presets/*.yaml"]
